op-deployer,ctb: Migrate DeployMIPS to new script API
