+++ packages/contracts-bedrock/scripts/deploy/DeployMIPS.s.sol
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.15;
+pragma solidity 0.8.15;
 
 // Forge
 import { Script } from "forge-std/Script.sol";
 
-// Scripts
-import { BaseDeployIO } from "scripts/deploy/BaseDeployIO.sol";
+// Libraries
 import { DeployUtils } from "scripts/libraries/DeployUtils.sol";
 
 // Interfaces
@@ -14,93 +13,52 @@ import { IMIPS } from "interfaces/cannon/IMIPS.sol";
 import { IMIPS2 } from "interfaces/cannon/IMIPS2.sol";
 import { StandardConstants } from "scripts/deploy/StandardConstants.sol";
 
-/// @title DeployMIPSInput
-contract DeployMIPSInput is BaseDeployIO {
-    // Specify the PreimageOracle to use
-    address internal _preimageOracle;
-
-    // Specify which MIPS version to use.
-    uint256 internal _mipsVersion;
-
-    function set(bytes4 _sel, uint256 _value) public {
-        if (_sel == this.mipsVersion.selector) {
-            require(_value == StandardConstants.MIPS_VERSION, "DeployMIPS: unsupported mips version");
-            _mipsVersion = _value;
-        } else {
-            revert("DeployMIPS: unknown selector");
-        }
-    }
-
-    function set(bytes4 _sel, address _value) public {
-        if (_sel == this.preimageOracle.selector) {
-            require(_value != address(0), "DeployMIPS: preimageOracle cannot be empty");
-            _preimageOracle = _value;
-        } else {
-            revert("DeployMIPS: unknown selector");
-        }
-    }
-
-    function mipsVersion() public view returns (uint256) {
-        require(_mipsVersion != 0, "DeployMIPS: mipsVersion not set");
-        require(_mipsVersion == StandardConstants.MIPS_VERSION, "DeployMIPS: unsupported mips version");
-        return _mipsVersion;
-    }
-
-    function preimageOracle() public view returns (address) {
-        require(_preimageOracle != address(0), "DeployMIPS: preimageOracle not set");
-        return _preimageOracle;
-    }
-}
-
-/// @title DeployMIPSOutput
-contract DeployMIPSOutput is BaseDeployIO {
-    IMIPS internal _mipsSingleton;
-
-    function set(bytes4 _sel, address _value) public {
-        if (_sel == this.mipsSingleton.selector) {
-            require(_value != address(0), "DeployMIPS: mipsSingleton cannot be zero address");
-            _mipsSingleton = IMIPS(_value);
-        } else {
-            revert("DeployMIPS: unknown selector");
-        }
-    }
-
-    function mipsSingleton() public view returns (IMIPS) {
-        DeployUtils.assertValidContractAddress(address(_mipsSingleton));
-        return _mipsSingleton;
-    }
-}
-
 /// @title DeployMIPS
 contract DeployMIPS is Script {
-    function run(DeployMIPSInput _mi, DeployMIPSOutput _mo) public {
-        deployMipsSingleton(_mi, _mo);
-        assertValidDeploy(_mi, _mo);
+    struct Input {
+        // Specify the PreimageOracle to use
+        IPreimageOracle preimageOracle;
+        // Specify which MIPS version to use.
+        uint256 mipsVersion;
     }
 
-    function deployMipsSingleton(DeployMIPSInput _mi, DeployMIPSOutput _mo) internal {
-        uint256 mipsVersion = _mi.mipsVersion();
-        IPreimageOracle preimageOracle = IPreimageOracle(_mi.preimageOracle());
+    struct Output {
+        IMIPS mipsSingleton;
+    }
+
+    function run(Input memory _input) public returns (Output memory output_) {
+        assertValidInput(_input);
+
+        deployMipsSingleton(_input, output_);
+
+        assertValidOutput(_input, output_);
+    }
+
+    function deployMipsSingleton(Input memory _input, Output memory _output) internal {
+        uint256 mipsVersion = _input.mipsVersion;
 
         IMIPS singleton = IMIPS(
             DeployUtils.createDeterministic({
                 _name: "MIPS64",
-                _args: DeployUtils.encodeConstructor(abi.encodeCall(IMIPS2.__constructor__, (preimageOracle, mipsVersion))),
+                _args: DeployUtils.encodeConstructor(
+                    abi.encodeCall(IMIPS2.__constructor__, (_input.preimageOracle, mipsVersion))
+                ),
                 _salt: DeployUtils.DEFAULT_SALT
             })
         );
 
         vm.label(address(singleton), "MIPSSingleton");
-        _mo.set(_mo.mipsSingleton.selector, address(singleton));
+        _output.mipsSingleton = singleton;
     }
 
-    function assertValidDeploy(DeployMIPSInput _mi, DeployMIPSOutput _mo) public view {
-        DeployUtils.assertValidContractAddress(address(_mo.mipsSingleton()));
-        assertValidMipsSingleton(_mi, _mo);
+    function assertValidInput(Input memory _input) public pure {
+        require(address(_input.preimageOracle) != address(0), "DeployMIPS: preimageOracle not set");
+        require(_input.mipsVersion != 0, "DeployMIPS: mipsVersion not set");
+        require(_input.mipsVersion == StandardConstants.MIPS_VERSION, "DeployMIPS: unsupported mips version");
     }
 
-    function assertValidMipsSingleton(DeployMIPSInput _mi, DeployMIPSOutput _mo) internal view {
-        IMIPS mips = _mo.mipsSingleton();
-        require(address(mips.oracle()) == address(_mi.preimageOracle()), "MIPS-10");
+    function assertValidOutput(Input memory _input, Output memory _output) public view {
+        DeployUtils.assertValidContractAddress(address(_output.mipsSingleton));
+        require(address(_output.mipsSingleton.oracle()) == address(_input.preimageOracle), "MIPS-10");
     }
 }
